6-size.c v.55

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long int y;
 	long long int z;
 	float f;
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(x));
-printf("Size of a long int : %lu byte(s)\n", (unsigned long)sizeof(y));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(z));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %zu byte(s)\n", sizeof(c));
+printf("Size of an int: %zu byte(s)\n", sizeof(x));
+printf("Size of a long int : %zu byte(s)\n", sizeof(y));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(z));
+printf("Size of a float: %zu byte(s)\n", sizeof(f));
 return (0);
 }
